Extract Canvas trait. Close #11
